+++ src/components/DevtoolsPane/index.tsx
@@ -27,6 +27,7 @@ import { DataChannelForm } from './DataChannelForm';
 import { DataChannelsForm } from './DataChannelsForm';
 import { DisconnectButton } from './DisconnectButton';
 import { DisplayResolutionForm } from './DisplayResolutionForm';
+import { DisposeMediaButton } from './DisposeMediaButton';
 import { E2EEForm } from './E2EEForm';
 import { EchoCancellationForm } from './EchoCancellationForm';
 import { EchoCancellationTypeForm } from './EchoCancellationTypeForm';
@@ -43,6 +44,7 @@ import { MultistreamForm } from './MultistreamForm';
 import { NoiseSuppressionForm } from './NoiseSuppressionForm';
 import { ReconnectForm } from './ReconnectForm';
 import { ReloadDevicesButton } from './ReloadDevicesButton';
+import { RequestMediaButton } from './RequestMediaButton';
 import { ResizeModeForm } from './ResizeModeForm';
 import { ResolutionForm } from './ResolutionForm';
 import { RoleForm } from './RoleForm';
@@ -481,6 +483,12 @@ export const DevtoolsPane: React.FC = () => {
       <RowMediaDevices />
       <hr className="hr-form" />
       <div className="row">
+        {role !== 'recvonly' ? (
+          <>
+            <RequestMediaButton />
+            <DisposeMediaButton />
+          </>
+        ) : null}
         <ConnectButton />
         <DisconnectButton />
         <StartRecordingButton />